Passcode entering and transition to next level
We made a text box where the player can enter the passcode, where if correct the game will take them into the next room or 'level'. We also added the correct art for the start screen and library, and made an error object for the passcode being wrong. However we need the error to show up only when the player enters the wrong code, so next time we will figure that out.

Co-Authored-By: marmif03 <147386849+marmif03@users.noreply.github.com>