Dragging objects
We coded the glasses, spoon and wig to be able to be dragged. Once dragged the sprites of the spoon and wig will change to make more sense, the glasses can stay the same. These 3 objects will also snap back in place when released, the sprites will also change back to the original.

Co-Authored-By: marmif03 <147386849+marmif03@users.noreply.github.com>