Canteen Scene Completed and Ending Scenes Completed
We started both the canteen and ending scenes and finished them in one session.

Created tasks for canteen (closed) which allows the player to 'transition' into canteen (open), and also later transition to the ending scenes which includes 3 rooms of narration.

We also added dialogues, sprites, objects, and fixes in each of these rooms (5 total)

Co-Authored-By: marmif03 <147386849+marmif03@users.noreply.github.com>